+++ devel/libevbase/src/evrtsig.h
@@ -11,7 +11,7 @@ int evrtsig_update(EVBASE *evbase, EVENT *event);
 /* Delete event from evbase */
 int evrtsig_del(EVBASE *evbase, EVENT *event);
 /* Loop evbase */
-int evrtsig_loop(EVBASE *evbase, short, struct timeval *tv);
+int evrtsig_loop(EVBASE *evbase, int, struct timeval *tv);
 /* Reset evbase */
 void evrtsig_reset(EVBASE *evbase);
 /* Clean evbase */
