+++ devel/libevbase/src/evwin32.c
@@ -21,7 +21,7 @@ int evwin32_del(EVBASE *evbase, EVENT *event)
     return -1;
 }
 /* Loop evbase */
-int evwin32_loop(EVBASE *evbase, short loop_flags, struct timeval *tv)
+int evwin32_loop(EVBASE *evbase, int loop_flags, struct timeval *tv)
 {
     return -1;
 }
